ai-service/requirements.txt
# 3D Chess AI Service - Python Dependencies

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.3

# Machine Learning (CPU-only for smaller image)
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.1.0
numpy>=1.24.0

# Chess
python-chess==1.999

# Utilities
python-dotenv==1.0.1
